+++ plugin/CHANGELOG.md
@@ -2,6 +2,9 @@ Change Log
 ==========
 
 ## Unreleased
+- JUnit 5.10.2
+- Raise minimum supported versions for AGP and Gradle to 8.0.x and 8.0, respectively
+- Allow overriding the version of the instrumentation libraries applied with the plugin
 
 ## 1.10.0.0 (2023-11-05)
 - JUnit 5.10.0